torch float32 everywhere

+++ ImputerExperiments/Locally_run_code/utils.py
@@ -355,7 +355,7 @@ def add_missing(X, add_missing = 0.05, missing_type = 'MAR'):
     missing_mask = missing_mask.mask(missing_mask.isna(), True)
     missing_mask = missing_mask.mask(missing_mask.notna(), False)
     X = X.mask(X.isna(), 0)
-    T = torch.tensor(X.to_numpy())
+    T = torch.tensor(X.to_numpy(), dtype=torch.float32)
 
     match missing_type:
         case 'MAR':
@@ -382,7 +382,7 @@ def add_missing(X, add_missing = 0.05, missing_type = 'MAR'):
 def MCAR(X, p_miss):
     out = {'X': X.double()}
     for p in p_miss: 
-        mask = (torch.rand(X.shape) < p).double()
+        mask = (torch.rand(X.shape, dtype=torch.float32) < p).double()
         X_nas = X.clone()
         X_nas[mask.bool()] = np.nan
         model_name = 'Missing'
@@ -395,25 +395,25 @@ def MAR(X,p_miss,p_obs=0.5):
     out = {'X': X.double()}
     for p in p_miss:
         n, d = X.shape
-        mask = torch.zeros(n, d).bool()
+        mask = torch.zeros(n, d, dtype=torch.float32).bool()
         num_no_missing = max(int(p_obs * d), 1)
         num_missing = d - num_no_missing
         obs_samples = np.random.choice(d, num_no_missing, replace=False)
         copy_samples = np.array([i for i in range(d) if i not in obs_samples])
         len_obs = len(obs_samples)
         len_na = len(copy_samples)
-        coeffs = torch.randn(len_obs, len_na).double()
+        coeffs = torch.randn(len_obs, len_na, dtype=torch.float32).double()
         Wx = X[:, obs_samples].mm(coeffs)
-        coeffs /= torch.std(Wx, 0, keepdim=True)
+        coeffs /= torch.std(Wx, 0, keepdim=True, dtype=torch.float32)
         coeffs.double()
         len_obs, len_na = coeffs.shape
-        intercepts = torch.zeros(len_na)
+        intercepts = torch.zeros(len_na, dtype=torch.float32)
         for j in range(len_na):
             def f(x):
-                return torch.sigmoid(X[:, obs_samples].mv(coeffs[:, j]) + x).mean().item() - p
+                return torch.sigmoid(X[:, obs_samples].mv(coeffs[:, j]) + x, dtype=torch.float32).mean().item() - p
             intercepts[j] = optimize.bisect(f, -50, 50)
-        ps = torch.sigmoid(X[:, obs_samples].mm(coeffs) + intercepts)
-        ber = torch.rand(n, len_na)
+        ps = torch.sigmoid(X[:, obs_samples].mm(coeffs) + intercepts, dtype=torch.float32)
+        ber = torch.rand(n, len_na, dtype=torch.float32)
         mask[:, copy_samples] = ber < ps
         X_nas = X.clone()
         X_nas[mask.bool()] = np.nan
@@ -451,10 +451,10 @@ def MNAR_mask_logistic(X, p_miss, p_params =.5, exclude_inputs=True):
     out = {'X_init_MNAR': X.double()}
     for p in p_miss: 
         n, d = X.shape
-        to_torch = torch.is_tensor(X) ## output a pytorch tensor, or a numpy array
+        to_torch = torch.is_tensor(X, dtype=torch.float32) ## output a pytorch tensor, or a numpy array
         if not to_torch:
-            X = torch.from_numpy(X)
-        mask = torch.zeros(n, d).bool() if to_torch else np.zeros((n, d)).astype(bool)
+            X = torch.from_numpy(X, dtype=torch.float32)
+        mask = torch.zeros(n, d, dtype=torch.float32).bool() if to_torch else np.zeros((n, d)).astype(bool)
         d_params = max(int(p_params * d), 1) if exclude_inputs else d ## number of variables used as inputs (at least 1)
         d_na = d - d_params if exclude_inputs else d ## number of variables masked with the logistic model
         ### Sample variables that will be parameters for the logistic regression:
@@ -465,25 +465,25 @@ def MNAR_mask_logistic(X, p_miss, p_params =.5, exclude_inputs=True):
         ### Pick coefficients so that W^Tx has unit variance (avoids shrinking)
         len_obs = len(idxs_params)
         len_na = len(idxs_nas)
-        coeffs = torch.randn(len_obs, len_na).double()
+        coeffs = torch.randn(len_obs, len_na, dtype=torch.float32).double()
         Wx = X[:, idxs_params].mm(coeffs)
-        coeffs /= torch.std(Wx, 0, keepdim=True)
+        coeffs /= torch.std(Wx, 0, keepdim=True, dtype=torch.float32)
         coeffs.double()
         ### Pick the intercepts to have a desired amount of missing values
         len_obs, len_na = coeffs.shape
-        intercepts = torch.zeros(len_na)
+        intercepts = torch.zeros(len_na, dtype=torch.float32)
         for j in range(len_na):
             def f(x):
-                return torch.sigmoid(X[:, idxs_params].mv(coeffs[:, j]) + x).mean().item() - p
+                return torch.sigmoid(X[:, idxs_params].mv(coeffs[:, j]) + x, dtype=torch.float32).mean().item() - p
             intercepts[j] = optimize.bisect(f, -50, 50)
-        ps = torch.sigmoid(X[:, idxs_params].mm(coeffs) + intercepts)
-        ber = torch.rand(n, d_na)
+        ps = torch.sigmoid(X[:, idxs_params].mm(coeffs) + intercepts, dtype=torch.float32)
+        ber = torch.rand(n, d_na, dtype=torch.float32)
         mask[:, idxs_nas] = ber < ps
         ## If the inputs of the logistic model are excluded from MNAR missingness,
         ## mask some values used in the logistic model at random.
         ## This makes the missingness of other variables potentially dependent on masked values
         if exclude_inputs:
-            mask[:, idxs_params] = torch.rand(n, d_params) < p
+            mask[:, idxs_params] = torch.rand(n, d_params, dtype=torch.float32) < p
         X_nas = X.clone()
         X_nas[mask.bool()] = np.nan
         model_name = 'Missing'
